code editor UI/openai-python-backend/requirements.txt
black==23.3.0
flask==2.2.5
openai==0.28.1
python-dotenv==1.0.0

